pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fptuchatbot"
version = "1.0.0"
description = "Production-ready RAG Chatbot for FPT University with Hybrid Retrieval"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "FPT University AI Team"}
]
keywords = ["rag", "chatbot", "vietnamese", "retrieval", "gemini", "faiss", "bm25"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    
    # PDF Processing
    "pymupdf>=1.23.0",
    "unstructured[pdf]>=0.12.0",
    "unstructured-inference>=0.7.0",
    "pillow>=10.2.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "camelot-py[cv]>=0.11.0",
    "pdfplumber>=0.10.3",
    
    # Embeddings & Vector Store
    "sentence-transformers>=2.3.0",
    "faiss-cpu>=1.7.4",
    "rank-bm25>=0.2.2",
    
    # LLM
    "google-generativeai>=0.3.0",
    
    # NLP & Text Processing
    "tiktoken>=0.5.2",
    "nltk>=3.8.1",
    "underthesea>=6.7.0",
    
    # Utils
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "orjson>=3.9.0",
    "xxhash>=3.4.0",
    "datasketch>=1.6.0",
    
    # Monitoring
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
    "ipdb>=0.13.13",
]

gpu = [
    "faiss-gpu>=1.7.4",
]

[project.scripts]
fptuchatbot-ingest = "scripts.ingest:main"
fptuchatbot-search = "scripts.search:main"
fptuchatbot-server = "scripts.run_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fptuchatbot*", "scripts*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=fptuchatbot --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

